backend/requirements.txt
Django==5.2.7
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
Pillow
